@@ -2,11 +2,11 @@ package com.clexi.clexi.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.ActionBar;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import androidx.appcompat.app.ActionBar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.ContextMenu;
 import android.view.Menu;
@@ -18,7 +18,6 @@ import com.clexi.clexi.app.Consts;
 import com.clexi.clexi.bluetoothle.BleManager;
 import com.clexi.clexi.bluetoothle.object.ApduCommand;
 import com.clexi.clexi.dialog.ConfirmDialog;
-import com.clexi.clexi.dialog.FlyingFOB;
 import com.clexi.clexi.model.access.DbManager;
 import com.clexi.clexi.model.object.Account;
 import com.clexi.clexi.test.TestCommands;
